+++ project-canvas/src/components/backlog-view/helpers/draggingHelpers.ts
